requirements.txt
pandas
sqlalchemy
psycopg2-binary
dash
plotly
python-dotenv